Refactored the whole logic, trying to make imports work... Stil not, asking for reviews!
