+++ app/models/users.py
@@ -7,3 +7,10 @@ class User(BaseModel):
         self.email = email
         self.first_name = first_name
         self.last_name = last_name
+
+    def user_dict(self):
+        return {
+            'email': self.email,
+            'first_name': self.first_name,
+            'last_name': self.last_name
+        }
