+++ app/persistence/ipersistencemanager.py
